requirements.txt
Pillow==12.1.1
numpy==2.4.3
pandas==3.0.1
matplotlib==3.10.8
seaborn==0.13.2
tqdm==4.67.3
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
ipython==9.1.0
ipykernel==6.30.0
jupyter==1.1.1
notebook==7.4.0
ipywidgets==8.1.8
